+++ NEWS.md
@@ -4,6 +4,8 @@
 * Change `ctdDecimate()` to add a new `na.rm` argument (issue #2192).
 * Change `plot.argo()` to handle `which="sigma0 profile"` (issue #2184).
 * Change `plot.section()` to handle bottom traces better (issue #2194).
+* Change `plot.section()` to show data as points properly (issue #2195).
+* Change `read.adp.rdi()` to handle large (> 2.147 GB) files (issue #2196).
 * Change `resizableLabel()` to handle `item=`"N"`. Also, document partial-match
   behaviour (issue #2197).
 * Change `sunAngle()` to handle vector arguments better (issue #2178).